HARMONY-1442: More code cleanup and fix tests to make sure mocha -w works everywhere.
